Create a Rust CI workflow for just building cc_bindings_from_rs
cargo test doesn't work yet (fails to build!), and I'd rather start iterating on the workflow with a subset that does work. So, the idea is, create the workflow, and then enable it and make sure to surface its results internally as well, and then fix coverage and make it run tests and so on.