Refactor using generation.
Previously we relied on visiting a `use` statement to generate a `using`
statement for an alias. Now we rely on `public_paths_by_def_id` to
determine aliases of a definition. Any path that isn't the canonical
path turns into an alias. This removes another usage of HIR. Our
previous logic relied on UsePath from the HIR node for a use statement
to generate the using clause.

PiperOrigin-RevId: 826215303