Split out determining unqualified name.
We have cases where we want the unqualified name of definitions that
have no canonical name. Right now we have a workaround in
`symbol_canonical_name` to support this. By splitting out the logic to
determine unqualified name we can remove that workaround and directly
ask for the unqualified name of symbols.

PiperOrigin-RevId: 827486177