Refactor ffi_11 crate to use newtypes.
Previously when we could be confident a C++ and Rust type matched in bit
width, it would be a type alias in ffi_11. Relying on aliases is
problematic for the rmeta interface (aliases don't exist at the MIR
level). Instead, use wrapper structs so our ffi types are considered
distinct at the MIR level.

PiperOrigin-RevId: 830498336