`ctor`: make `Reconstruct` a safe trait, and provide it for all types.
The documentation above explicitly says that this is safe for all Rust types,
and we are after all implementing this for Rust types. There is other language
about base class subobjects and `[[no_unique_address]]` members in C++, but:

*   Again, this is a Rust trait for Rust types. One would hope that Crubit isn't
    projecting C++ types to Rust types that don't actually work like Rust types.
*   Neither "is a base class subobject" or "has no unique address" are
    attributes of *types*, only particular objects. A trait can't help
    distinguish them.
*   Even if they were in the type system, they have nothing to do with `!Unpin`
    as far as I can tell. So this code already would have been wrong.

This leaves the code in a weird state, where we have a trait implemented for all
objects (and no negative implementations). But that was already the case for
`ReconstructUnchecked`; I'm not really sure what the author's intent was
originally. I'll leave that for the Crubit team to sort out, but for now this
unblocks me when needing to reconstruct a `!Unpin` type.

PiperOrigin-RevId: 832466808