Change the check for whether to collect evidence for a Return and Arg->Params to hasInferable in addition to isSupportedPointerType.
For now, we keep the isSupportedPointerType, since the code will then proceed to do `getPointerValue(ReturnExpr, ...)`, which could crash if `hasInferable` but not `isSupportedPointerType` (e.g., the currently behind a flag support for vector<T*>).

Checking 'hasInferable' matches what the caller would check for gatherInferableSlots, and what countInferableSlots would do.

PiperOrigin-RevId: 833402835