Add support for c_*longlong in ffi_11.
It's an open question how to handle long long since it overlaps with
long on some platforms. I've opted to make longlong/ulonglong a wrapper
type on platforms where it overlaps with long. `std::int64_t` is the
same type as long on these platforms, so for our tests to behave as
expected it's important longlong(/ulonglong) is a new type that will be
spelled "long long" (and not std::int64_t).

I've tried to match the original intent, but there was some talk of
using isize/usize that I don't follow. So possible I'm missing important
nuance here.

NB: I added support for `#[repr(transparent)]` to improve support for
wrapper types in this module.
PiperOrigin-RevId: 834274409