Correctly handle default constructor for #[repr(transparent)] structs.
Previously, we assumed all structs were initialized by reference and
always passed a `this` pointer to our default constructor. This is not
the case for `#[repr(transparent)]` structs which are passed by value.
Add a check for that case and change our ctor to set `this` using the
returned value rather than pass it by ref.

PiperOrigin-RevId: 835316870