Nullability (internal): include note message directly in the diagnostic.
It seems more natural to include the details of the note when the diagnostic is created, as is done for the note's source range, rather than obtaining it later from a case analysis on the diagnostic code.

This also means we don't need to create a new error code whenever we want variations in the associated note.

PiperOrigin-RevId: 835345053