Move UniformReprTemplateType identification logic to C++
Previously, we would eagerly import all template type arguments and put them in the `template_specialization` field of `ir::Record`. As described in b/465422519, this leads to incorrect behavior because it doesn't take into account that clang won't complete an instantiation until it needs to, leading us to sometimes cache a type as incomplete when it might be completable.

The fix is to not eagerly import these types unless we _know_ they are used, which happens in the select cases that `UniformReprTemplateType` is used (unique_ptr, vector, and certain bridge types).

PiperOrigin-RevId: 839250494