Require a Drop impl for CRUBIT_OWNED_POINTER types
When using the OWNED_POINTER/OWNED_POINTEE feature, we should encourage the
implementation of a Drop method on the Rust wrapper type to help ensure that
the underlying C++ is properly freed when the wrapping type is dropped.

To do this, we unconditionally implement Drop for the generated wrapper type,
and then delegate the call to a DropImpl function; it's the users
responsibility to implement this method and include it in an
"additional_rust_srcs" file.

PiperOrigin-RevId: 839436338