Fix enum emission under `custom_ffi_types`.
Enums emit constants that construct instances from numeric literals. This fails for enums that are backed by an `ffi_11` type that is a newtype and not a typealias (notably `c_char`). It does not suffice to use the `Into` implementation on `c_char` in the generated bindings because we need our solution to be const. This CL creates a `new` method that is const and uses that in the generated bindings.

PiperOrigin-RevId: 840427730