Delete bazel CI support.
GitHub Actions is better! Publicly viewable, etc.

We don't actually _have_ a github actions CI thing set up to test the bazel build -- but, well, we don't have a working bazel-CI at the moment, either. (Focus is temporarily on cargo setups, which are more useful to most direct users atm.) This is a bit on faith that it'll be possible to re-set-up the bazel test suite with GitHub Actions. At any rate, the bazelci config is dead code, and if we do want to revive it, we ca always undelete it.

PiperOrigin-RevId: 842854297