Update implementation of `stable_def_id_cmp`.
Previously, this logic relied on spans to order DefIds, which appears to
be non-deterministic in practice. This CL updates our ordering to solely
rely on DefPathHash which is intended to be stable across compilation
sessions.

I'm not totally sure why the Spans were non-deterministic. There's a
couple spots where it looks like they get interned and the possibility
of observing insertion order might sneak in. I did not diagnose further,
since def path hashes fixed the issue.

PiperOrigin-RevId: 844946550