Turn enable HIR types feature flag into a noop.
With the removal of HIR types, we could eventually remove SugaredTy as
everywhere we construct it now uses `missing_hir`. I've left that as
future work to avoid too much refactoring in the CL that removes HIR
types.

Some of the tests were broken and it didn't appear obvious how to update them to allow depending on ffi_11, so I've removed them in favor of relying on primitives types test.

PiperOrigin-RevId: 845392864