Update stable_def_id_cmp to use printed def path.
DefPathHash is stable within one build context, but changes with
compiler version (which we do often), presenting problems for the
stability of our godlen tests. Switch to use full printed def paths for
ordering definitions. This shouldn't rely on compiler version or
metadata, and also shouldn't encounter the instability of `Span`.

PiperOrigin-RevId: 845788004