Format incomplete records without type tags when requested.
Inside of other types, they could be formatted as e.g. `struct Foo;`, when a complete struct would be just `Foo`. This is because of a really fiddly diverging code path that basically just seems to be a bad factorization. In practice this seems to result in no changes to the output, and is a pure simplification.

PiperOrigin-RevId: 847964078