Use fully qualified names for error messages.
This, in turn, gives us a useful and distinct key to aggregate on with error reporting infra.

Also, it disambiguates errors in the event that there's many things with the same name. ;)

---

The reason this works is that record IDs are always recorded as the enclosing item ID. Indeed, the record_id field is completely redundant -- this CL shows that much, and I will delete it in a followup CL.

PiperOrigin-RevId: 849309009