Delete redundant record ID in the IR.
A function _never_ has a record ID different from its enclosing record. And it isn't even useful in a typed way, because the record_id might not be a Item::Record! May as well always use the enclosing_record_id (and check for Namespace if you really need to).

The previous CLs in the chain relied on the fact that this was redundant. Deleting it properly helps for the proof.

PiperOrigin-RevId: 849311347