Ignore `final` function specifier in Crubit.
This can impact bindings for a final destructor, but we check that via checking if the type itself is "effectively" final (either final or has a final destructor). At the level of functions, we don't need to know whether the function can be overridden, because we don't implement inheritance yet.

PiperOrigin-RevId: 850242570